LeetCode/src/johnny/algorithm/leetcode/Solution206.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package johnny.algorithm.leetcode;
import johnny.algorithm.leetcode.common.ListNode;

/**
 * Reverse Linked List.
 * Reverse a singly linked list.
 * @author Johnny
 */
public class Solution206 {
    public ListNode reverseList(ListNode head) {
        ListNode prev = null;
        while (head != null) {
            ListNode next = head.next;
            head.next = prev;
            prev = head;
            head = next;
        }
        return prev;
    }
}
